feat(tests): add limited mock quizzes and refactor quiz mocks for improved test coverage
- Introduced a new `limitedMockQuizzes.ts` file to provide a shorter quiz list for specific test cases.
- Refactored `mockQuizzes.ts` to utilize external JSON data for quizzes, enhancing maintainability.
- Created `mockQuizzesData.json` to store quiz data, streamlining the mock setup for tests.
